src/test/java/pe/edu/pucp/johannmorales/thesis/algorithms/genetic/model/BitArrayTest.java
package pe.edu.pucp.johannmorales.thesis.algorithms.genetic.model;

import static org.assertj.core.api.Assertions.assertThat;

import org.junit.Test;

public class BitArrayTest {

  @Test
  public void flip_ShouldSetTrue() {
    assertThat(BitArray.create("101").flip(1)).isEqualTo(BitArray.create("111"));
    assertThat(BitArray.create("011").flip(2)).isEqualTo(BitArray.create("111"));
  }

  @Test
  public void flip_ShouldSetFalse() {
    assertThat(BitArray.create("010").flip(1)).isEqualTo(BitArray.create("000"));
    assertThat(BitArray.create("100").flip(2)).isEqualTo(BitArray.create("000"));
  }

  @Test
  public void setFrom_ShouldSetTrueFromStart() {
    assertThat(BitArray.create("000").setFrom(0, true)).isEqualTo(BitArray.create("111"));
    assertThat(BitArray.create("100").setFrom(0, true)).isEqualTo(BitArray.create("111"));
  }

  @Test
  public void setFrom_ShouldSetFalseFromStart() {
    assertThat(BitArray.create("111").setFrom(0, false)).isEqualTo(BitArray.create("000"));
    assertThat(BitArray.create("011").setFrom(0, false)).isEqualTo(BitArray.create("000"));
  }

  @Test
  public void setFrom_ShouldSetTrueFromPosition() {
    assertThat(BitArray.create("000").setFrom(1, true)).isEqualTo(BitArray.create("110"));
    assertThat(BitArray.create("010").setFrom(1, true)).isEqualTo(BitArray.create("110"));
  }

  @Test
  public void setFrom_ShouldSetFalseFromPosition() {
    assertThat(BitArray.create("111").setFrom(1, false)).isEqualTo(BitArray.create("001"));
    assertThat(BitArray.create("101").setFrom(1, false)).isEqualTo(BitArray.create("001"));
  }

  @Test
  public void setTo_ShouldSetTrueToEnd() {
    assertThat(BitArray.create("000").setTo(3, true)).isEqualTo(BitArray.create("111"));
    assertThat(BitArray.create("100").setTo(3, true)).isEqualTo(BitArray.create("111"));
  }

  @Test
  public void setTo_ShouldSetFalseToEnd() {
    assertThat(BitArray.create("111").setTo(3, false)).isEqualTo(BitArray.create("000"));
    assertThat(BitArray.create("011").setTo(3, false)).isEqualTo(BitArray.create("000"));
  }

  @Test
  public void setTo_ShouldSetTrueToPosition() {
    assertThat(BitArray.create("000").setTo(2, true)).isEqualTo(BitArray.create("011"));
    assertThat(BitArray.create("010").setTo(2, true)).isEqualTo(BitArray.create("011"));
  }

  @Test
  public void setTo_ShouldSetFalseToPosition() {
    assertThat(BitArray.create("111").setTo(2, false)).isEqualTo(BitArray.create("100"));
    assertThat(BitArray.create("101").setTo(2, false)).isEqualTo(BitArray.create("100"));
  }

}